hotfix(develop): 'page' folder renamed to 'components'; 'services' folder and api file created
