+++ client/package.json
@@ -8,7 +8,6 @@
     "@testing-library/user-event": "^13.2.1",
     "@types/jest": "^27.0.1",
     "@types/node": "^16.7.13",
-    "@types/react": "^18.0.0",
     "@types/react-dom": "^18.0.0",
     "axios": "^0.27.2",
     "react": "^18.1.0",
@@ -43,6 +42,7 @@
     ]
   },
   "devDependencies": {
+    "@types/react": "^18.0.8",
     "@types/styled-components": "^5.1.25"
   }
 }
