+++ client/src/App.tsx
@@ -4,7 +4,7 @@ import { theme } from './styles/theme';
 import { GlobalStyle } from './styles/global';
 
 
-import { Home } from './pages';
+import { Home } from './components';
 
 function App() {
   return (
