+++ client/src/services/api-axios.ts
@@ -1,0 +1,7 @@
+import axios from "axios";
+
+const apiAxios = axios.create({
+  baseURL: "http://localhost:3001/",
+});
+
+export default apiAxios;
